requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pika==1.1.0
psycopg2-binary==2.8.3
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
PyJWT==1.7.1
drf-yasg==1.16.1
packaging==19.1
django-filter==2.2.0
django-cors-headers==3.1.0
drf-extensions==0.5.0
pika==1.1.0
xhtml2pdf==0.2.3
redis==3.3.8
django-request-logging==0.7.0
